chapter_001/src/main/java/ru/job4j/loop/Paint.java
package ru.job4j.loop;

/**
 * Paint - рисование пирамиды в псевдографике
 *
 * @author Ksenya Kaysheva (murrcha@me.com)
 * @version $Id$
 * @since 0.1
 */
public class Paint {

    /**
     * Method pyramid - рисует пирамиду в псевдографике заданной высоты
      * @param height - высота пирамиды
     * @return - строка в форме пирамиды
     */
    public String pyramid(int height) {
        StringBuilder screen = new StringBuilder();
        int weight = 2 * height - 1;
        for (int row = 0; row != height; row++) {
            for (int column = 0; column != weight; column++) {
                if ((row >= height - column - 1) && (row + height - 1 >= column)) {
                    screen.append("^");
                } else {
                    screen.append(" ");
                }
            }
            screen.append(System.lineSeparator());
        }
        return screen.toString();
    }

    /**
     * Method rightTrl - рисует правую половину пирамиды
     * @param height - высота пирамиды
     * @return - строка в форме правой половины пирамиды
     */
    public String rightTrl(int height) {
        StringBuilder screen = new StringBuilder();
        int weight = height;
        for (int row = 0; row != height; row++) {
            for (int column = 0; column != weight; column++) {
                if (row >= column) {
                    screen.append("^");
                } else {
                    screen.append(" ");
                }
            }
            screen.append(System.lineSeparator());
        }
        return screen.toString();
    }

    /**
     * Method leftTrl - рисует левую половину пирамиды
     * @param height - высота пирамиды
     * @return - строка в форме левой половины пирамиды
     */
    public String leftTrl(int height) {
        StringBuilder screen = new StringBuilder();
        int weight = height;
        for (int row = 0; row != height; row++) {
            for (int column = 0; column != weight; column++) {
                if (row >= weight - column - 1) {
                    screen.append("^");
                } else {
                    screen.append(" ");
                }
            }
            screen.append(System.lineSeparator());
        }
        return screen.toString();
    }
}
